+++ packages/rx-react/src/index.ts
@@ -1,13 +1,21 @@
 /**
  * @since 1.0.0
  */
+/* eslint-disable @typescript-eslint/no-empty-object-type */
+import * as Reactive from "@effect-rx/rx/Reactive"
 import * as Registry from "@effect-rx/rx/Registry"
 import * as Result from "@effect-rx/rx/Result"
 import * as Rx from "@effect-rx/rx/Rx"
 import type * as RxRef from "@effect-rx/rx/RxRef"
+import { Runtime } from "effect"
 import * as Cause from "effect/Cause"
+import * as Context from "effect/Context"
+import * as Effect from "effect/Effect"
 import type * as Exit from "effect/Exit"
+import { constNull } from "effect/Function"
 import { globalValue } from "effect/GlobalValue"
+import * as Layer from "effect/Layer"
+import type { Scope } from "effect/Scope"
 import * as React from "react"
 import * as Scheduler from "scheduler"
 
@@ -31,6 +39,16 @@ export * as Rx from "@effect-rx/rx/Rx"
  * @category modules
  */
 export * as RxRef from "@effect-rx/rx/RxRef"
+/**
+ * @since 1.0.0
+ * @category modules
+ */
+export * as Reactive from "@effect-rx/rx/Reactive"
+/**
+ * @since 1.0.0
+ * @category modules
+ */
+export * as ReactiveRef from "@effect-rx/rx/ReactiveRef"
 
 /**
  * @since 1.0.0
@@ -349,3 +367,205 @@ export const useRxRefProp = <A, K extends keyof A>(ref: RxRef.RxRef<A>, prop: K)
  */
 export const useRxRefPropValue = <A, K extends keyof A>(ref: RxRef.RxRef<A>, prop: K): A[K] =>
   useRxRef(useRxRefProp(ref, prop))
+
+/**
+ * @since 1.0.0
+ * @category Reactive
+ */
+export interface ReactiveComponent<Props extends Record<string, any>, R = never> {
+  (
+    props: Props & [R] extends [never] ? {
+        readonly context?: Context.Context<never> | undefined
+      } :
+      { readonly context: Context.Context<R> }
+  ): React.ReactNode
+
+  provide<AL, EL, RL>(layer: Layer.Layer<AL, EL, RL>): ReactiveComponent<Props, Exclude<R, AL> | RL>
+
+  render: Effect.Effect<(props: Props) => React.ReactNode, never, R | Reactive.Reactive>
+}
+
+const depsAreEqual = (a: ReadonlyArray<any>, b: ReadonlyArray<any>) => {
+  if (a.length !== b.length) {
+    return false
+  }
+  for (let i = 0; i < a.length; i++) {
+    if (!Object.is(a[i], b[i])) {
+      return false
+    }
+  }
+  return true
+}
+
+const makeReactiveComponent = <Props extends Record<string, any>, E, R, Context = void>(options: {
+  readonly name: string
+  readonly build: (
+    props: Props,
+    context: Context
+  ) => Effect.Effect<React.ReactNode, E, R>
+  readonly layer: Layer.Layer<never>
+  readonly fallback: (props: Props) => React.ReactNode
+  readonly deps: (props: Props) => ReadonlyArray<any>
+  readonly beforeBuild?: (props: Props) => Context
+}): ReactiveComponent<Props, R> => {
+  function ReactiveComponent(props: Props & { readonly context?: Context.Context<R> }) {
+    const ref = React.useRef<{
+      readonly subscribable: Reactive.Subscribable<React.ReactNode, E>
+      previousDeps: ReadonlyArray<any>
+      cancel: () => void
+      timeout: number | undefined
+      result: Result.Result<React.ReactNode, E>
+      updateCount: number
+      rerender: () => void
+    }>(undefined as any)
+
+    const deps = options.deps(props)
+    const context = options.beforeBuild?.(props) as Context
+
+    if (ref.current === undefined || !depsAreEqual(ref.current.previousDeps, deps)) {
+      if (ref.current) {
+        ref.current.cancel()
+        if (ref.current.timeout !== undefined) {
+          clearTimeout(ref.current.timeout)
+          ref.current.timeout = undefined
+        }
+      }
+      const layer = props.context
+        ? Layer.provideMerge(options.layer, Layer.succeedContext(props.context))
+        : options.layer
+      const subscribable = Reactive.toSubscribable(layer)(
+        options.build(props, context) as Effect.Effect<React.ReactNode, E, Reactive.Reactive>
+      )
+      ref.current = ref.current ?
+        {
+          ...ref.current,
+          subscribable,
+          previousDeps: deps
+        } :
+        {
+          subscribable,
+          previousDeps: deps,
+          timeout: undefined,
+          result: Result.initial(true),
+          updateCount: 0,
+          rerender() {
+            ref.current.updateCount++
+          }
+        } as any
+      ref.current.cancel = subscribable.subscribe((result) => {
+        ref.current.result = result
+        ref.current.rerender()
+      })
+    }
+
+    const [count, setCount] = React.useState<number>(ref.current.updateCount)
+
+    React.useEffect(() => {
+      ref.current.rerender = () => setCount(++ref.current.updateCount)
+      if (ref.current.updateCount !== count) {
+        setCount(ref.current.updateCount)
+      }
+      if (ref.current.timeout !== undefined) {
+        clearTimeout(ref.current.timeout)
+        ref.current.timeout = undefined
+      }
+      return () => {
+        ref.current.timeout = setTimeout(ref.current.cancel, 100)
+      }
+    }, [])
+
+    if (ref.current.result._tag === "Initial") {
+      return options.fallback(props)
+    } else if (ref.current.result._tag === "Failure") {
+      throw Cause.squash(ref.current.result.cause)
+    }
+    return ref.current.result.value
+  }
+  ReactiveComponent.displayName = options.name
+  ReactiveComponent.provide = function provide(layer: Layer.Layer<any, any, any>) {
+    return makeReactiveComponent({
+      ...options,
+      layer: options.layer === Layer.empty ? layer : Layer.provideMerge(options.layer, layer) as any
+    })
+  }
+  let renderCache: WeakMap<Reactive.Reactive["Type"], (props: Props) => React.ReactNode> | undefined
+  ReactiveComponent.render = Effect.contextWith((context: Context.Context<any>) => {
+    const reactive = Context.unsafeGet(context, Reactive.Reactive)
+    if (renderCache?.has(reactive)) {
+      return renderCache.get(reactive)!
+    }
+    function Wrapped(props: Props) {
+      return React.createElement(ReactiveComponent, { ...props, context })
+    }
+    Wrapped.displayName = `${options.name}.render`
+    if (renderCache === undefined) {
+      renderCache = new WeakMap()
+    }
+    renderCache.set(reactive, Wrapped)
+    return Wrapped
+  })
+  return ReactiveComponent as any
+}
+
+const defaultDeps = (props: Record<string, any>) => Object.values(props)
+
+/**
+ * @since 1.0.0
+ * @category Reactive
+ */
+export const component = <Props extends Record<string, any> = {}>(name: string): {
+  <E, R>(
+    build: (props: Props) => Effect.Effect<React.ReactNode, E, R>,
+    options?: {
+      readonly fallback?: (props: Props) => React.ReactNode
+      readonly deps?: (props: Props) => ReadonlyArray<any>
+    }
+  ): ReactiveComponent<Props, Exclude<R, Reactive.Reactive | Scope>>
+  <E, R, Context>(
+    beforeBuild: (props: Props) => Context,
+    build: (props: Props, context: Context) => Effect.Effect<React.ReactNode, E, R>,
+    options?: {
+      readonly fallback?: (props: Props) => React.ReactNode
+      readonly deps?: (props: Props) => ReadonlyArray<any>
+    }
+  ): ReactiveComponent<Props, Exclude<R, Reactive.Reactive | Scope>>
+} =>
+  function() {
+    let beforeBuild: any
+    let build: any
+    let options: any
+    if (typeof arguments[1] === "function") {
+      beforeBuild = arguments[0]
+      build = arguments[1]
+      options = arguments[2]
+    } else {
+      build = arguments[0]
+      options = arguments[1]
+    }
+    return makeReactiveComponent({
+      name,
+      build,
+      layer: Layer.empty,
+      fallback: options?.fallback ?? constNull,
+      deps: options?.deps ?? defaultDeps,
+      beforeBuild
+    }) as any
+  }
+
+/**
+ * @since 1.0.0
+ * @category Reactive
+ */
+export const action = <Args extends ReadonlyArray<any>, A, E, R>(
+  f: (...args: Args) => Effect.Effect<A, E, R>
+): Effect.Effect<(...args: Args) => Promise<A>, never, R> =>
+  Effect.withFiberRuntime((parent) =>
+    Effect.succeed((...args: Args) => {
+      const runtime = Runtime.make({
+        context: parent.currentContext as Context.Context<R>,
+        fiberRefs: parent.getFiberRefs(),
+        runtimeFlags: Runtime.defaultRuntime.runtimeFlags
+      })
+      return Runtime.runPromise(runtime)(f(...args))
+    })
+  )