+++ packages/rx-vue/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @effect-rx/rx-vue
 
+## 0.11.13
+
+### Patch Changes
+
+- Updated dependencies [[`fc4e4d4`](https://github.com/tim-smart/effect-rx/commit/fc4e4d4a524b62bcb814cd78f953cbf9f84ce984)]:
+  - @effect-rx/rx@0.34.13
+
+## 0.11.12
+
+### Patch Changes
+
+- Updated dependencies [[`1c7562e`](https://github.com/tim-smart/effect-rx/commit/1c7562ec05d1da02c5deb51b869038554932d143)]:
+  - @effect-rx/rx@0.34.12
+
+## 0.11.11
+
+### Patch Changes
+
+- Updated dependencies [[`b62af9c`](https://github.com/tim-smart/effect-rx/commit/b62af9c0d195edf620d00a7a51552b18180e9a3d)]:
+  - @effect-rx/rx@0.34.11
+
+## 0.11.10
+
+### Patch Changes
+
+- Updated dependencies [[`44ddfc2`](https://github.com/tim-smart/effect-rx/commit/44ddfc27d0cec928531bec4a5760efc1e473a9b7)]:
+  - @effect-rx/rx@0.34.10
+
 ## 0.11.9
 
 ### Patch Changes